worker/requirements.txt
aiogram>=3.21.0
python-telegram-bot>=20.0
asyncpg>=0.30.0
python-dotenv>=1.0.0 